# Speed up VMDB::Util.log_duration_gz