Basically the idea is to read the file once (previous version did it twice). It doesn't give much performance gain against small files, but with big files there is a noticeable win.

The behavior changed a bit: before this commit only first and last 50 lines were taken into account. Now the method greps through all the lines in the log. Even despite this fact, the performance is not worse than in the previous version. Benchmarks:

Big file (750MB+ unarchived, 38MB gzipped, 2.8M lines)

```
2.2.3 :037 > t = Time.now; VMDB::Util   .get_log_start_end_times('log/big.log.gz'); Time.now - t
 => 34.867957
2.2.3 :038 > t = Time.now; VMDB::UtilOld.get_log_start_end_times('log/big.log.gz'); Time.now - t
 => 41.215036
```

35s vs 41s - looks like a win. I ran it several times, every time the new version was under 35s and the old one was over 41s.

Medium-sized file (5.4MB, 426KB gzipped, 32434 lines)
```ruby
Benchmark.ips do |x|
  x.config(time: 50, warmup: 10)
  x.report('after')  { VMDB::Util   .get_log_start_end_times('log/medium.log.gz') }
  x.report('before') { VMDB::UtilOld.get_log_start_end_times('log/medium.log.gz') }
  x.compare!
end
__END__
Calculating -------------------------------------
               after     1.000  i/100ms
              before     1.000  i/100ms
-------------------------------------------------
               after      3.124  (± 0.0%) i/s -    157.000
              before      3.010  (± 0.0%) i/s -    151.000
Comparison:
               after:        3.1 i/s
              before:        3.0 i/s - 1.04x slower
```

Smallish file (28K, 4.2K gzipped, 100 lines)

```
Benchmark.ips do |x|
  x.report('after')  { VMDB::Util   .get_log_start_end_times('log/small.log.gz') }
  x.report('before') { VMDB::UtilOld.get_log_start_end_times('log/small.log.gz') }
  x.compare!
end
__END__
Calculating -------------------------------------
               after     9.000  i/100ms
              before     9.000  i/100ms
-------------------------------------------------
               after    103.052  (± 2.9%) i/s -    522.000
              before     99.273  (± 2.0%) i/s -    504.000

Comparison:
               after:      103.1 i/s
              before:       99.3 i/s - 1.04x slower
```

Stats related to medium- and small-sized jump around (sometimes the old version outperforms the new version by a percent or two), so the win only affects big log files. On my laptop, at least.

I also made several refactorings in order to be able to introduce this change.