@chessbyte 2015-10-26
@gtanzillo @jrafanie Please review